refactor(nyaa): split 1303-line file into mod + scraper; bump to 1.2.5
Phase 5 of the refactor plan + milestone bump.

  services/nyaa/
    mod.rs       308   public surface: HTTP_CLIENT, NYAA_BASE,
                        SearchResult / SearchOptions / SearchResponse types,
                        search() entry point, fetch_view_result,
                        enrich_results_with_group_map,
                        extract_hash / base32_decode_infohash
    scraper.rs  1012   all scraper internals: 8 DOM selectors, 4 regex
                        statics (BATCH_RANGE_RE, SEASON_MARKER_RE,
                        ROMAN_SEASON_MARKER_RE, SINGLE_EP_RE), parse_results,
                        ClassifiedFields + classify_search_result,
                        sanitize_query_for_nyaa, detect_batch, parse_size,
                        parse_int, parse_view_page + all tests

Three imports to note:
  - `use ::scraper::{Html, Selector};` (leading `::`) — the external crate
    name now collides with the sibling module `mod scraper`, so the bare
    path `scraper::Html` would resolve to `self::scraper` (this file)
    rather than the external crate.
  - `use super::{NYAA_BASE, SearchOptions, SearchResult, extract_hash};`
    in scraper.rs — scraper internals need these from mod.rs.
  - Test module imports enrich_results_with_group_map from super::super
    for the group-map enrichment tests.

No external caller changes. Phase 5 was marked optional in the plan
but completes the monolith hit-list — nothing over 2300 lines left.

Also in this commit: version bump 1.2.0 → 1.2.5 in Cargo.toml, with
Cargo.lock regenerated via `cargo build --offline`.

cargo fmt / clippy -D warnings / build --locked / test --locked all green
(601 tests).
